directly cut-paste tm code from rw
